Micro-optimizations, is now much faster than the jump table.

| Method |     Mean |    Error |   StdDev | Scaled | Allocated |
|------- |---------:|---------:|---------:|-------:|----------:|
|   Jump | 639.6 ns | 3.368 ns | 3.150 ns |   0.70 |       0 B |
|   Orig | 918.4 ns | 6.598 ns | 6.172 ns |   1.00 |       0 B |
|  Array | 494.4 ns | 3.768 ns | 3.525 ns |   0.54 |       0 B |